actions: Fix Windows Go version
Prior to this patch, the Go version for the Windows tests in Github
actions was set as `1.16.x`, which mysteriously resolved to
`1.15.15`[1].

With this patch, the Go version for the Windows tests is set as a semver
range.

This patch also bumps `actions/setup-go`[2] to the latest version,
consistent with the other platforms.

[1]: https://github.com/kubernetes-csi/csi-driver-nfs/runs/4211323920?check_suite_focus=true
[2]: https://github.com/actions/setup-go